refactor: 클래스 타입 수정 record >> class
