+++ api/store/src/main/java/com/backtothefuture/store/dto/response/ReservationResponseDto.java
@@ -1,11 +1,18 @@
 package com.backtothefuture.store.dto.response;
 
-import lombok.Builder;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 
-@Builder
-public record ReservationResponseDto(
-        String name, // 상품 이름
-        int count,  // 상품의 주문 수량
-        int price  // 주문에서 해당 상품의 총 가격
-) {
+@Getter
+@AllArgsConstructor
+public class ReservationResponseDto {
+
+    private String name; // 상품 이름
+    private int count;  // 상품의 주문 수량
+    private int price;  // 주문에서 해당 상품의 총 가격
+
+    public void calculateTotalPrice() {
+        this.price = price * count;
+    }
+
 }
